TheGkApp/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.2
django-ckeditor==5.7.1
django-formtools==2.1
django-js-asset==1.2.2
django-richtextfield==1.4.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
plotly==4.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.3
